requirements.txt
utm==0.7.0
Django==3.1.7
rest_condition==1.0.3
django_imagekit==4.0.2
djangorestframework_gis==0.17
drf_yasg==1.20.0
drf_autodocs==0.4.4
piexif==1.1.3
django_dirtyfields==1.5.0
django_filter==2.4.0
ExifRead==2.3.2
djangorestframework==3.12.2
Pillow==8.1.2
